+++ CHANGES.md
@@ -5,6 +5,8 @@
 ---------------------
 
 - fix Component.require_one raising a configuration error with strict=False
+- Adds command `./batou secrets decrypttostdout` to decrypt a secrets file to stdout.
+  - Useful for integration with git diff using textconv, see documentation for installation instructions.
 - The error message shown when a batou deployment fails to start due to another
   deployment running at the same time, now correctly informs the user that
   another deployment is running instead of showing a traceback. (#446)